+++ langgraph.json
@@ -4,7 +4,7 @@
         "pyproject.toml"
     ],
     "graphs": {
-        "agent": "./experimental/agents/create_react_agent/agent.py:analytics_agent"
+        "superstore": "./experimental/agents/superstore/agent.py:analytics_agent"
     },
     "env": ".env"
 }
